Refactor TeamExporter into ModelExporter
